upd: refactor to allow metric_filters to override manual check management
