+++ internal/check/check.go
@@ -6,6 +6,7 @@
 package check
 
 import (
+	"encoding/json"
 	"fmt"
 	"regexp"
 	"strings"
@@ -19,6 +20,12 @@ import (
 	"github.com/spf13/viper"
 )
 
+// initCheck initializes the check for the agent.
+// 1. fetch a check explicitly provided via CID
+// 2. search for a check matching the current system
+// 3. create a check for the system if --check-create specified
+// if fetched, found, or created - set Check.bundle
+// otherwise, return an error
 func (c *Check) initCheck(cid string, create bool) error {
 	var bundle *apiclient.CheckBundle
 
@@ -55,69 +62,69 @@ func (c *Check) initCheck(cid string, create bool) error {
 	return nil
 }
 
-func (c *Check) setCheck() error {
-	// retrieve the check via the Circonus API or create a new check (if configured to do so)
-	isCreate := viper.GetBool(config.KeyCheckCreate)
-	isManaged := viper.GetBool(config.KeyCheckEnableNewMetrics)
-	isReverse := viper.GetBool(config.KeyReverse)
-	cid := viper.GetString(config.KeyCheckBundleID)
-
-	var bundle *apiclient.CheckBundle
-
-	// if explicit cid configured, attempt to fetch check bundle using cid
-	if cid != "" {
-		b, err := c.fetchCheck(cid)
-		if err != nil {
-			return errors.Wrapf(err, "fetching check for cid %s", cid)
-		}
-		bundle = b
-	} else {
-		// if no cid configured, attempt to find check bundle matching this system
-		b, found, err := c.findCheck()
-		if err != nil {
-			if !isCreate || found != 0 {
-				return errors.Wrap(err, "unable to find a check for this system")
-			}
-			c.logger.Info().Msg("no existing check found, creating")
-			// attempt to create if not found and create flag ON
-			b, err = c.createCheck()
-			if err != nil {
-				return errors.Wrap(err, "creating new check for this system")
-			}
-		}
-		bundle = b
-	}
-
-	if bundle == nil {
-		return errors.New("invalid Check object state, bundle is nil")
-	}
-
-	c.bundle = bundle
-
-	if isManaged {
-		c.logger.Debug().Msg("setting metric states")
-		err := c.setMetricStates(&bundle.Metrics)
-		if err != nil {
-			return errors.Wrap(err, "setting metric states")
-		}
-	}
-
-	// the metrics from the reference bundle are not needed in memory
-	// as they will never be used again.
-	c.bundle.Metrics = []apiclient.CheckBundleMetric{}
-
-	if isReverse {
-		// populate reverse configuration
-		c.logger.Debug().Msg("setting reverse config")
-		err := c.setReverseConfig()
-		if err != nil {
-			return errors.Wrap(err, "setting up reverse configuration")
-		}
-	}
-	c.logger.Debug().Msg("done updating check")
-
-	return nil
-}
+// func (c *Check) setCheck() error {
+// 	// retrieve the check via the Circonus API or create a new check (if configured to do so)
+// 	isCreate := viper.GetBool(config.KeyCheckCreate)
+// 	isManaged := viper.GetBool(config.KeyCheckEnableNewMetrics)
+// 	isReverse := viper.GetBool(config.KeyReverse)
+// 	cid := viper.GetString(config.KeyCheckBundleID)
+//
+// 	var bundle *apiclient.CheckBundle
+//
+// 	// if explicit cid configured, attempt to fetch check bundle using cid
+// 	if cid != "" {
+// 		b, err := c.fetchCheck(cid)
+// 		if err != nil {
+// 			return errors.Wrapf(err, "fetching check for cid %s", cid)
+// 		}
+// 		bundle = b
+// 	} else {
+// 		// if no cid configured, attempt to find check bundle matching this system
+// 		b, found, err := c.findCheck()
+// 		if err != nil {
+// 			if !isCreate || found != 0 {
+// 				return errors.Wrap(err, "unable to find a check for this system")
+// 			}
+// 			c.logger.Info().Msg("no existing check found, creating")
+// 			// attempt to create if not found and create flag ON
+// 			b, err = c.createCheck()
+// 			if err != nil {
+// 				return errors.Wrap(err, "creating new check for this system")
+// 			}
+// 		}
+// 		bundle = b
+// 	}
+//
+// 	if bundle == nil {
+// 		return errors.New("invalid Check object state, bundle is nil")
+// 	}
+//
+// 	c.bundle = bundle
+//
+// 	if isManaged {
+// 		c.logger.Debug().Msg("setting metric states")
+// 		err := c.setMetricStates(&bundle.Metrics)
+// 		if err != nil {
+// 			return errors.Wrap(err, "setting metric states")
+// 		}
+// 	}
+//
+// 	// the metrics from the reference bundle are not needed in memory
+// 	// as they will never be used again.
+// 	c.bundle.Metrics = []apiclient.CheckBundleMetric{}
+//
+// 	if isReverse {
+// 		// populate reverse configuration
+// 		c.logger.Debug().Msg("setting reverse config")
+// 		err := c.setReverseConfig()
+// 		if err != nil {
+// 			return errors.Wrap(err, "setting up reverse configuration")
+// 		}
+// 	}
+// 	c.logger.Debug().Msg("done updating check")
+//
+// 	return nil
+// }
 
 func (c *Check) fetchCheck(cid string) (*apiclient.CheckBundle, error) {
 	if cid == "" {
@@ -227,10 +234,13 @@ func (c *Check) createCheck() (*apiclient.CheckBundle, error) {
 	cfg.Brokers = []string{brokerCID}
 
 	if viper.GetBool(config.KeyCheckEnableNewMetrics) {
+		cfg.MetricFilters = defaults.CheckMetricFilters
 		if viper.GetString(config.KeyCheckMetricFilters) != "" {
-			// parse metric filters
-		} else {
-			cfg.MetricFilters = defaults.CheckMetricFilters
+			var filters [][]string
+			if err := json.Unmarshal([]byte(viper.GetString(config.KeyCheckMetricFilters)), &filters); err != nil {
+				return nil, errors.Wrap(err, "parsing check metric filters")
+			}
+			cfg.MetricFilters = filters
 		}
 	}
